finance_system/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
sqlparse==0.5.5
asgiref==3.11.1
tzdata==2025.3